Refactor folder-structure, adjust to build-environment
