--- src/Algorithm/Algorithm.html
@@ -1,29 +1,0 @@
-<div class="form-group form-group-sm">
-    <label id="algo_label"
-        class="controls label label-default"
-        for="algo_select">algorithm</label>
-    <select id="algo_select"
-        class="btn btn-default"
-        type="list"
-        list="detail_list"
-        bind:value='algorithm'
-        ref:algorithm
-        on:click='fire( "select" )'
-    >
-    {{#each algorithms as al}}
-    <option value={{al.key}}>{{al.key}}</option>
-    {{/each}}
-    </select>
-</div>
-algorithm: {{algorithm}}
-<script>
-export default {
-  onrender () {
-    let algorithms = this.get('algorithms')
-    if (algorithms && algorithms[0] && algorithms[0].key) {
-      this.set({ algorithm: algorithms[0].key })
-      this.fire('select', algorithms[0].key)
-    }
-  }
-}
-</script>
