--- src/Algorithm/Neighbor.html
@@ -1,49 +1,0 @@
-<div class="form-group form-group-sm">
-    <label id="state_label"
-        class="controls label label-default"
-        for="nb_select">neighborhood</label>
-    <select id="nb_select"
-        class="btn btn-default"
-        type="list"
-        list="nb"
-        bind:value='neighborhood'
-        ref:neighborhood
-        on:click='fire( "select" )'
-    >
-    {{#if neighbors}}
-    {{#each neighbors as nb, i}}
-        {{#if i === 0}}
-        <option selected>{{nb}}</option>
-        {{else}}
-        <option>{{nb}}</option>
-        {{/if}}
-    {{/each}}
-    {{/if}}
-    </select>
-</div>
-<div>
-neighborhood: (
-{{ neighborhood }}
-)
-</div>
-<script>
-export default {
-  onrender () {
-    this.observe('neighbors', neighbors => {
-      if (this.get('neighbors')) {
-        console.log('observing neighbor types to be like ' + neighbors)
-        let nb = this.get('neighborhood')
-        if (!nb) {
-          this.set({ neighborhood: neighbors[0] })
-        } else if (nb) {
-          let m = Array.from(neighbors)
-          .map((n) => {
-            n
-          })
-          console.log(m)
-        }
-      }
-    }, { defer: true })
-  }
-}
-</script>
