+++ src/components/Algorithm.html
@@ -1,0 +1,57 @@
+<div>
+  <div class="form-group form-group-sm">
+    <label id="algo_label"
+        class="controls label label-default"
+        for="algo_select">algorithm</label>
+    <select id="algo_select"
+        class="btn btn-default"
+        bind:value="algorithm"
+    >
+      {{#if algorithms}}
+      {{#each algorithms as al}}
+      <option value={{al.key}}>{{al.key}}</option>
+      {{/each}}
+      {{/if}}
+      </select>
+  </div>
+  {{#if algorithms}}
+    {{#each algorithms as al}}
+      {{#if al}}
+      {{ al }}
+      {{/if}}
+    {{/each}}
+  {{/if}}
+  {{#if neighbors}}
+  <li>
+    <Neighbor
+      ref:neighborhood
+      neighbors="{{ neighbors }}"
+      on:click='fire( "select" )'/>
+  </li>
+  {{/if}}
+</div>
+<script>
+import { findKey } from 'lodash-es'
+
+import Neighbor from './Neighbor.html'
+
+function selectedAlgorithm (algorithms) {
+  return findKey(algorithms, (algorithm) => { return { key: algorithm } })
+}
+
+export default {
+  data () {
+    return {
+      algorithms: this.algorithms
+    }
+  },
+  onrender () {
+  },
+  computed: {
+    neighbors: algorithms => selectedAlgorithm(algorithms)
+  },
+  components: {
+    Neighbor
+  }
+}
+</script>
