+++ src/components/AlgorithmControls.html
@@ -1,0 +1,9 @@
+<Algorithm algorithms="{{ algorithms }}" />
+<script>
+import Algorithm from './Algorithm.html'
+
+export default {
+  components: { Algorithm }
+}
+</script>
+
