+++ src/components/DetailControl.html
@@ -18,7 +18,7 @@
 <script>
 export default {
   onrender () {
-    this.set({ detail: 4 })
+    // this.set({ detail: 3 })
   }
 }
 </script>
