+++ src/js/app.js
@@ -1,7 +1,7 @@
 
 // import * as _ from 'lodash'
 import settings from './settings.js'
-import Controls from './Controls.html'
+import Controls from '../components/Controls.html'
 
 export default class App {
   constructor () {
