+++ src/js/settings.js
