+++ src/main.js
@@ -1,5 +1,5 @@
 // Main program
 
-import App from './app.js'
+import App from './js/app.js'
 
 var app = new App() // eslint-disable-line no-unused-vars
